+++ cmd/record.go
@@ -16,8 +16,12 @@ limitations under the License.
 package cmd
 
 import (
+	"os"
+	"strings"
+
 	"github.com/google/test-server/internal/config"
 	"github.com/google/test-server/internal/record"
+	"github.com/google/test-server/internal/redact"
 	"github.com/spf13/cobra"
 )
 
@@ -33,7 +37,14 @@ target server, and all requests and responses will be recorded.`,
 		if err != nil {
 			panic(err)
 		}
-		err = record.Record(config, recordingDir)
+
+		secrets := os.Getenv("TEST_SERVER_SECRETS")
+		redactor, err := redact.NewRedact(strings.Split(secrets, ","))
+		if err != nil {
+			panic(err)
+		}
+
+		err = record.Record(config, recordingDir, redactor)
 		if err != nil {
 			panic(err)
 		}